testando algoritmo

+++ test/dijkstra_gfg.c
@@ -33,8 +33,8 @@ void printSolution(int dist[]) {
 
 // Function that implements Dijkstra's single source shortest path algorithm
 // for a graph represented using adjacency matrix representation
-void dijkstra(int graph[V][V], int src) {
-    int dist[V]; // The output array. dist[i] will hold the shortest
+void dijkstra(float graph[V][V], int src) {
+    float dist[V]; // The output array. dist[i] will hold the shortest
     // distance from src to i
 
     bool sptSet[V]; // sptSet[i] will be true if vertex i is included in shortest
@@ -74,15 +74,23 @@ void dijkstra(int graph[V][V], int src) {
 int main() {
     /* Let us create the example graph discussed above */
 
-    int graph[V][V] = {
-        {0, 4, 0, 0, 0, 0, 0, 8, 0},  {4, 0, 8, 0, 0, 0, 0, 11, 0}, {0, 8, 0, 7, 0, 4, 0, 0, 2},
-        {0, 0, 7, 0, 9, 14, 0, 0, 0}, {0, 0, 0, 9, 0, 10, 0, 0, 0}, {0, 0, 4, 14, 10, 0, 2, 0, 0},
-        {0, 0, 0, 0, 0, 2, 0, 1, 6},  {8, 11, 0, 0, 0, 0, 1, 0, 7}, {0, 0, 2, 0, 0, 0, 6, 7, 0}};
+    float graph[V][V] = {
+        {0, 4.0, 0, 0, 0, 0, 0, 8, 0},  {4.0, 0, 8.0, 0, 0, 0, 0, 11, 0}, {0, 8, 0, 7, 0, 4, 0, 0, 2},
+        {0, 0, 7.0, 0, 9, 14, 0, 0, 0}, {0, 0, 0, 9.0, 0, 10, 0, 0, 0}, {0, 0, 4, 14, 10, 0, 2, 0, 0},
+        {0, 0, 0, 0, 0, 2, 0, 1, 6.0},  {8, 11, 0, 0, 0, 0, 1, 0, 7}, {0, 0, 2, 0, 0, 0, 6, 7, 0}};
 
-    
     dijkstra(graph, 0);
-
-
 
     return 0;
 }
+
+// Vertex           Distance from Source
+// 0                0
+// 1                4
+// 2                12
+// 3                19
+// 4                21
+// 5                11
+// 6                9
+// 7                8
+// 8                14
